feat(teamStats): Split team stats into two tabs
