+++ static/app/views/organizationStats/header.tsx
@@ -10,7 +10,7 @@ import {Organization} from 'sentry/types';
 
 type Props = {
   organization: Organization;
-  activeTab: 'stats' | 'team';
+  activeTab: 'stats' | 'issues' | 'health';
 };
 
 function StatsHeader({organization, activeTab}: Props) {
@@ -20,7 +20,7 @@ function StatsHeader({organization, activeTab}: Props) {
         <StyledLayoutTitle>{t('Stats')}</StyledLayoutTitle>
       </Layout.HeaderContent>
       <Layout.HeaderActions>
-        {activeTab === 'team' && (
+        {activeTab !== 'stats' && (
           <Button
             title={t('Send us feedback via email')}
             size="small"
@@ -36,9 +36,15 @@ function StatsHeader({organization, activeTab}: Props) {
             {t('Usage Stats')}
           </Link>
         </li>
-        <li className={`${activeTab === 'team' ? 'active' : ''}`}>
-          <Link to={`/organizations/${organization.slug}/stats/team/`}>
-            {t('Team Stats')}
+        <li className={`${activeTab === 'issues' ? 'active' : ''}`}>
+          <Link to={`/organizations/${organization.slug}/stats/issues/`}>
+            {t('Issues')}
+            <FeatureBadge type="beta" />
+          </Link>
+        </li>
+        <li className={`${activeTab === 'health' ? 'active' : ''}`}>
+          <Link to={`/organizations/${organization.slug}/stats/health/`}>
+            {t('Health')}
             <FeatureBadge type="beta" />
           </Link>
         </li>
