+++ static/app/views/organizationStats/teamInsights/index.tsx
@@ -2,8 +2,6 @@ import {cloneElement, isValidElement} from 'react';
 
 import Feature from 'sentry/components/acl/feature';
 import NoProjectMessage from 'sentry/components/noProjectMessage';
-import SentryDocumentTitle from 'sentry/components/sentryDocumentTitle';
-import {t} from 'sentry/locale';
 import {Organization} from 'sentry/types';
 import withOrganization from 'sentry/utils/withOrganization';
 
@@ -16,13 +14,11 @@ function TeamInsightsContainer({children, organization}: Props) {
   return (
     <Feature organization={organization} features={['team-insights']}>
       <NoProjectMessage organization={organization}>
-        <SentryDocumentTitle title={t('Project Reports')} orgSlug={organization.slug}>
-          {children && isValidElement(children)
-            ? cloneElement(children, {
-                organization,
-              })
-            : (children as React.ReactChild)}
-        </SentryDocumentTitle>
+        {children && isValidElement(children)
+          ? cloneElement(children, {
+              organization,
+            })
+          : (children as React.ReactChild)}
       </NoProjectMessage>
     </Feature>
   );
